[TASK] wait_for_db() now checks logs to determine if database is up
Now we check the db log to determine if database is truly up & running.
Previously we tried to connect and run `status` command inside MySQL.
That usually worked, but sporadically we've had issues when the next
client connection from the following commands (secure_and_tidy_db,
show_db_status, create_admin_user) was refused. That resulted in
the container failed to start, as the error occurred in the main run.sh
script.